PMacc simulation control: separate into hpp and cpp files
Move implementations into cpp files.
