+++ include/pmacc/simulationControl/TimeInterval.hpp
@@ -25,6 +25,7 @@
 
 #include <chrono>
 #include <iostream>
+#include <iomanip>
 #include <sstream>
 #include <string>
